✨ feat: storing chatid even after closing browser
re-creates new chat id when closing the chat bubble
